nav bar + image upload in cloudinary

+++ controllers/employees.controller.js
@@ -1,3 +1,6 @@
+const upload = require('../config/multer')
+const cloudinary = require('../config/cloudinary')
+
 const express = require('express')
 const router = express.Router()
 const isSignedIn = require('../middleware/is-signed-in')
@@ -11,13 +14,17 @@ router.get('/new', (req, res) => {
 
 
 // post the new employee to the db
-router.post('/', isSignedIn, async (req, res) => {
+router.post('/', isSignedIn,  upload.single('image'), async (req, res) => {
   try {
     req.body.user = req.session.user._id
+    req.body.image = {
+      url: req.file.path,
+      cloudinary_id: req.file.filename
+    }
     const createdEmployee = await Employee.create(req.body)
-
+    
     res.redirect('/employees')
-
+    
   } catch (error) {
     console.log(error)
     res.send('Something went wrong in posting an employee')
@@ -28,8 +35,8 @@ router.post('/', isSignedIn, async (req, res) => {
 //  getting all employees from db and rendering them in employees.index.ejs
 router.get('/', async (req, res) => {
   try {
-    const foundEmployees = await Employee.find()
-
+    const foundEmployees = await Employee.find({user: req.session.user._id})
+    
     res.render('employees/index.ejs', {foundEmployees: foundEmployees})
 
   } catch (error) {
@@ -43,6 +50,10 @@ router.get('/', async (req, res) => {
 router.get('/:employeeId', async (req, res) => {
   try {
     const foundEmployee = await Employee.findById(req.params.employeeId).populate('user')
+     
+    if (!foundEmployee.user._id.equals(req.session.user._id)) {
+      return res.redirect(`/employees`)
+    }
 
     res.render('employees/show.ejs', { foundEmployee })
 
@@ -57,7 +68,15 @@ router.get('/:employeeId', async (req, res) => {
 router.delete('/:employeeId', isSignedIn, async (req, res) => {
   try {
     const foundEmployee = await Employee.findById(req.params.employeeId).populate('user')
-    
+ 
+    if (!foundEmployee.user._id.equals(req.session.user._id)) {
+      return res.redirect(`/employees`)
+    }
+
+    if (foundEmployee.image?.cloudinary_id) {
+      await cloudinary.uploader.destroy(foundEmployee.image.cloudinary_id)
+    }
+
     if(foundEmployee.user._id.equals(req.session.user._id)) {
       await foundEmployee.deleteOne()
       return res.redirect('/employees')
@@ -75,7 +94,12 @@ router.delete('/:employeeId', isSignedIn, async (req, res) => {
 router.get('/:employeeId/edit', isSignedIn, async (req, res) => {
   try {
     const foundEmployee = await Employee.findById(req.params.employeeId).populate('user')
-    
+
+     
+    if (!foundEmployee.user._id.equals(req.session.user._id)) {
+      return res.redirect(`/employees`)
+    }
+
     if(foundEmployee.user._id.equals(req.session.user._id)) {
       return res.render('employees/edit.ejs', { foundEmployee: foundEmployee })
     }
@@ -93,10 +117,26 @@ router.put('/:employeeId', isSignedIn, async (req, res) => {
   try {
     const foundEmployee = await Employee.findById(req.params.employeeId).populate('user')
     
-    if (foundEmployee.user._id.equals(req.session.user._id)) {
-      const updatedEmployee = await Employee.findByIdAndUpdate(req.params.employeeId, req.body, { new: true })
-      return res.redirect(`/employees/${req.params.employeeId}`)
+    if (!foundEmployee.user._id.equals(req.session.user._id)) {
+      return res.redirect(`/employees`)
     }
+
+    if (foundEmployee.user._id.equals(req.session.user._id)){ 
+      if (req.file && foundEmployee.image?.cloudinary_id) {
+        await cloudinary.uploader.destroy(foundEmployee.image.cloudinary_id)
+        foundEmployee.image.url = req.file.path
+        foundEmployee.image.cloudinary_id = req.file.filename
+      }
+
+      foundEmployee.name = req.body.name
+      foundEmployee.position = req.body.position
+      foundEmployee.salary = req.body.salary
+
+      await foundEmployee.save()
+      return res.redirect(`/employees/${req.params.employeeId}`)
+      
+    }
+
     return res.send('Not Authorized')
 
   } catch (error) {
